% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_txt_files_inmet.R
\name{read_txt_file_inmet}
\alias{read_txt_file_inmet}
\title{Read raw ASCII data file of INMET's automatic stations}
\usage{
read_txt_file_inmet(.file, .verbose = TRUE, .only.problems = FALSE)
}
\arguments{
\item{.file}{path to file}

\item{.verbose}{logical scalar}

\item{.only.problems}{logical scalar. TRUE to report problems, 
FALSE default to return meteorological data}
}
\value{
a tibble with data or problems
}
\description{
Read raw ASCII data file of INMET's automatic stations
}
