% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_txt_files_inmet.R
\name{import_txt_files_inmet}
\alias{import_txt_files_inmet}
\title{Import raw data files of automatic stations}
\usage{
import_txt_files_inmet(files, verbose = TRUE, only.problems = FALSE)
}
\arguments{
\item{files}{character vector with path to files}

\item{verbose}{logical scalar. If TRUE, print messages and warnings.}

\item{only.problems}{logical scalar. Use TRUE to return a tibble 
with problems information on file and FALSE to return meteorological data.}
}
\value{
a tibble
}
\description{
Import raw data files of automatic stations
}
\examples{

library(dplyr); library(purrr); library(stringr)
# missing columns problem example
myfile <- system.file("extdata", "A838.txt", package = "inmetwrangler")
myfile
 A838_problems <- import_txt_files_inmet(files = myfile, 
                                         verbose = TRUE, 
                                         only.problems = TRUE)
 A838_data <- import_txt_files_inmet(files = myfile, 
                                     verbose = TRUE, 
                                     only.problems = FALSE)
#  View(slice(A838_data, A838_problems$row))

}
