% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-txt-files-inmet.R
\name{import_txt_files_inmet}
\alias{import_txt_files_inmet}
\title{Import raw data files of automatic stations}
\usage{
import_txt_files_inmet(files, verbose = TRUE, only.problems = FALSE)
}
\arguments{
\item{files}{character vecto with path to files}

\item{verbose}{logical scalar}

\item{only.problems}{logical scalar. Use TRUE to return a tibble 
with problems information on file and FALSE to return meteorological data.}
}
\value{
a tibble
}
\description{
Import raw data files of automatic stations
}
\examples{
library(dplyr); library(purrr); library(stringr)
txt_files <- list.files("RAW-DATA-8o-DISME",
                        recursive = TRUE, 
                        full.names = TRUE) \%>\%
   discard(., ~str_detect(.x, "\\\\.zip")) \%>\%
   keep(., ~str_detect(.x, "A838|A852"))
# missing columns problem example
 A838_problems <- import_txt_files_inmet(txt_files[1], 
                                         verbose = TRUE, 
                                         only.problems = TRUE)
 A838_data <- import_txt_files_inmet(txt_files[1], 
                                     verbose = TRUE, 
                                     only.problems = FALSE)
 View(slice(A838_data, A838_problems$row))
# too much columns problem example
# repeat for txt_files[2]
}
