% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_txt_files_inmet.R
\name{import_txt_files_inmet}
\alias{import_txt_files_inmet}
\title{Import raw data files of automatic stations}
\usage{
import_txt_files_inmet(files, verbose = TRUE, only.problems = FALSE,
  full.names = FALSE)
}
\arguments{
\item{files}{character vector with path to files}

\item{verbose}{logical scalar. If TRUE, print messages and warnings.}

\item{only.problems}{logical value. Set TRUE to return a tibble 
with problems information on file and FALSE to return meteorological data.}

\item{full.names}{a logical value. If TRUE, the directory path is prepended 
to the file names to give a relative file path. If FALSE, the file names 
(rather than paths) are returned.}
}
\value{
A data frame with one row for each problem and four columns:
  \item{row,col}{Row and column of problem}
  \item{expected}{What readr expected to find}
  \item{actual}{What it actually got}
  \item{file}{file name or the path to file (if \code{full.names} is TRUE)}
}
\description{
Import raw data files of automatic stations
}
\examples{

library(dplyr); library(purrr); library(stringr); library(readr)
# missing columns problem example
myfile <- system.file("extdata", "A838.txt", package = "inmetwrangler")
myfile
 A838_problems <- import_txt_files_inmet(files = myfile, 
                                         verbose = TRUE, 
                                         only.problems = TRUE)
 A838_data <- import_txt_files_inmet(files = myfile, 
                                     verbose = TRUE, 
                                     only.problems = FALSE)
#looking at rows
for(irow in A838_problems$row) read_lines(myfile, skip = irow-2, n_max = irow+2)
# View(slice(A838_data, A838_problems$row)) #  columns filled with NAs
}
