% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_txt_files_inmet.R
\name{import_txt_files_inmet}
\alias{import_txt_files_inmet}
\title{Import raw data files of automatic stations}
\usage{
import_txt_files_inmet(files, verbose = TRUE, only.problems = FALSE)
}
\arguments{
\item{files}{character vecto with path to files}

\item{verbose}{logical scalar}

\item{only.problems}{logical scalar. Use TRUE to return a tibble 
with problems information on file and FALSE to return meteorological data.}
}
\value{
a tibble
}
\description{
Import raw data files of automatic stations
}
\examples{
library(dplyr); library(purrr); library(stringr)
# missing columns problem example
myfile <- system.file("extdata", "A838.txt", package = "inmetwrangler")
myfile
}
